requirements.txt
Django==1.7.4
django-auth-ldap==1.2.0
django-extensions==1.5.0
django-ldapdb==0.3.2
djangorestframework==3.0.5
dj-static==0.0.6
Fabric==1.10.1
mysqlclient==1.3.5
passlib==1.6.2
python-ldap==2.4.15
